Empty buffer bugfix on replies
When receiving an empty transfer buffer in store_urb_bulk or store_urb_control_transfer_ex, get_buf properly returns a NULL transfer buffer pointer which is interpreted as a failure to allocate the memory. Instead, if the buffer is empty, just bypass the buffer copying/handling code.
Not necessary in store_urb_class_vendor but neater.